b_dynamodb_common_test/requirements.txt
 # Used for testing in AWS environment.
b-aws-testing-framework>=0.0.24,<1.0.0
# Used to create a DynamoDB table to test against.
# (Used only for integration tests).
aws-cdk.aws-dynamodb
# Used for sending requests.
urllib3>=1.25.10,<2.0.0
# Used for running tests with coverage.
pytest-cov>=2.0.0,<4.0.0
# HTTP mocking tool.
pook>=1.0.1,<2.0.0
# AWS SDK for python for invoking lambda functions.
boto3>=1.16.0,<2.0.0
